Skip broken vendor symlink rather than returning an error.

+++ internal/gps/strip_vendor.go
@@ -22,11 +22,7 @@ func stripVendor(path string, info os.FileInfo, err error) error {
 		}
 
 		if (info.Mode() & os.ModeSymlink) != 0 {
-			realInfo, err := os.Stat(path)
-			if err != nil {
-				return err
-			}
-			if realInfo.IsDir() {
+			if realInfo, err := os.Stat(path); err == nil && realInfo.IsDir() {
 				return os.Remove(path)
 			}
 		}
